refactor: divide the component
